Fold edit into profile page.
