--- ui/src/app/views/profile-edit/component.css
@@ -1,33 +1,0 @@
-h2 {
-  display: inline-block;
-}
-
-.field-wrapper {
-  margin: 0.5rem 0 1.5rem 0;
-}
-
-label.long-text-field {
-  display: block;
-}
-
-label clr-icon {
-  margin-left: 0.25rem;
-}
-
-input[type='text'] {
-  display: block;
-  width: 10rem;
-}
-
-div.field {
-  display: inline-block;
-  margin-right: 2rem;
-}
-input[type=text]:not([readonly]) {
-  padding: 0.75rem;
-  width: 14rem;
-}
-
-label {
-  padding-bottom: 0.5rem;
-}
