--- ui/src/app/views/profile-edit/component.html
@@ -1,90 +1,0 @@
-<h2>Profile</h2>
-<button (mouseover)="editHover=true" (mouseleave)="editHover=false"
-    (click)="editing = true" class="btn btn-icon">
-  <app-edit-icon [editHover]="editHover"></app-edit-icon>
-</button>
-<span *ngIf="!profileLoaded;else publicInfo" class="spinner spinner-sm"></span>
-<div class="alert alert-danger" *ngIf="errorText">
-  <div class="alert-items">
-    <div class="alert-item static">
-      <div class="alert-icon-wrapper">
-        <clr-icon class="alert-icon" shape="exclamation-circle"></clr-icon>
-      </div>
-      <span class="alert-text">{{errorText}}</span>
-    </div>
-  </div>
-</div>
-<ng-template #publicInfo>
-  <div class="static info">
-    <clr-icon shape="user" size="18" class="is-solid"></clr-icon> Shows on your public profile.
-  </div>
-</ng-template>
-<div class="field-wrapper">
-  <div class="field">
-    <label for="givenName" class="long-text-field">Given Name:<clr-icon shape="user" size="18" class="is-solid"></clr-icon></label>
-    <input *ngIf="profileLoaded else loading" type="text" autofocus
-      [(ngModel)]="profile.givenName" id="givenName" [disabled]="!editing">
-  </div>
-  <div class="field">
-    <label for="familyName" class="long-text-field">Family Name:<clr-icon shape="user" size="18" class="is-solid"></clr-icon></label>
-    <input *ngIf="profileLoaded else loading" type="text"
-      [(ngModel)]="profile.familyName" id="familyName" [disabled]="!editing">
-  </div>
-</div>
-<div class="field-wrapper">
-  <div class="field">
-    <label for="contactEmail" class="long-text-field">Contact Email:</label>
-    <input *ngIf="profileLoaded else loading" type="text"
-      [(ngModel)]="profile.contactEmail" id="contactEmail" [disabled]="!editing">
-  </div>
-</div>
-<h5>Username:</h5>
-<div *ngIf="profileLoaded else loading">{{profile.username}}</div>
-<ng-template #loading>
-  <div class="loading-area">
-    <span class="loading-text">Loading...</span>
-  </div>
-</ng-template>
-<div>
-  <h3>Identity Verification</h3>
-  <div *ngIf="!profileLoaded; else verifiedStatusMessage">
-    <!-- TODO(dmohs): Spinner. -->
-    Checking verification status...
-  </div>
-  <ng-template #verifiedStatusMessage>
-    <div *ngIf="profile.blockscoreIdVerificationStatus === 'UNVERIFIED'">
-      <div class="alert alert-warning">
-        <div class="alert-items">
-          <div class="alert-item static">
-            <div class="alert-icon-wrapper">
-              <clr-icon class="alert-icon" shape="exclamation-circle"></clr-icon>
-            </div>
-            <span class="alert-text">Your identity has not been verified.</span>
-          </div>
-        </div>
-      </div>
-    </div>
-    <div class="alert alert-success" *ngIf="profile.blockscoreIdVerificationStatus === 'VERIFIED'">
-      <div class="alert-items">
-        <div class="alert-item static">
-          <div class="alert-icon-wrapper">
-            <clr-icon class="alert-icon" shape="check-circle"></clr-icon>
-          </div>
-          <span class="alert-text">Your identity has been verified.</span>
-        </div>
-      </div>
-    </div>
-    <div class="alert alert-danger" *ngIf="profile.blockscoreIdVerificationStatus === 'REJECTED'">
-      <div class="alert-items">
-        <div class="alert-item static">
-          <div class="alert-icon-wrapper">
-            <clr-icon class="alert-icon" shape="exclamation-circle"></clr-icon>
-          </div>
-          <span class="alert-text">Your identity cannot be validated.</span>
-        </div>
-      </div>
-    </div>
-  </ng-template>
-  <button (click)="submitChanges()" class="btn">Update Profile</button>
-  <button (click)="editing=false" class="btn">Cancel</button>
-</div>
