+++ ui/src/app/views/profile-page/component.html
@@ -1,28 +1,46 @@
 <h2>Profile</h2>
 <button (mouseover)="editHover=true" (mouseleave)="editHover=false"
-    routerLink="edit" class="btn btn-icon">
+    (click)="editing = true" class="btn btn-icon">
   <app-edit-icon [editHover]="editHover"></app-edit-icon>
 </button>
-<span *ngIf="!profileLoaded" class="spinner spinner-sm"></span>
+<span *ngIf="!profileLoaded;else publicInfo" class="spinner spinner-sm"></span>
+<div class="alert alert-danger" *ngIf="errorText">
+  <div class="alert-items">
+    <div class="alert-item static">
+      <div class="alert-icon-wrapper">
+        <clr-icon class="alert-icon" shape="exclamation-circle"></clr-icon>
+      </div>
+      <span class="alert-text">{{errorText}}</span>
+    </div>
+  </div>
+</div>
+<ng-template #publicInfo>
+  <div class="static info">
+    <clr-icon shape="user" size="18" class="is-solid"></clr-icon> Shows on your public profile.
+  </div>
+</ng-template>
 <div class="field-wrapper">
   <div class="field">
-    <h5 class="field-label">Given Name:</h5>
-    <div *ngIf="profileLoaded else loading" class="data">{{profile.givenName}}</div>
+    <label for="givenName" class="long-text-field">Given Name:<clr-icon shape="user" size="18" class="is-solid"></clr-icon></label>
+    <input *ngIf="profileLoaded else loading" type="text" autofocus
+      [(ngModel)]="profile.givenName" id="givenName" [disabled]="!editing">
   </div>
   <div class="field">
-    <h5 class="field-label">Family Name:</h5>
-    <div *ngIf="profileLoaded else loading" class="data">{{profile.familyName}}</div>
+    <label for="familyName" class="long-text-field">Family Name:<clr-icon shape="user" size="18" class="is-solid"></clr-icon></label>
+    <input *ngIf="profileLoaded else loading" type="text"
+      [(ngModel)]="profile.familyName" id="familyName" [disabled]="!editing">
   </div>
 </div>
 <div class="field-wrapper">
   <div class="field">
-    <h5 class="field-label">Contact Email:</h5>
-    <div *ngIf="profileLoaded else loading" class="data">{{profile.contactEmail}}</div>
-    <div *ngIf="profileLoaded">
+    <label for="contactEmail" class="long-text-field">Contact Email:</label>
+    <div style="display: flex; justify-content: left" *ngIf="profileLoaded else loading">
+      <input type="text"
+        [(ngModel)]="profile.contactEmail" id="contactEmail" [disabled]="!editing"
+        style="margin-right: 1rem;">
       <div class="alert alert-warning"*ngIf="profile.emailVerificationStatus !== 'subscribed'">
         <div class="alert-items">
           <div class="alert-item static">
-            <div class="alert-icon-wrapper"><clr-icon class="alert-icon" shape="exclamation-circle"></clr-icon></div>
             <span class="alert-text">Your email address is awaiting verification. Please check your inbox for an email from All of Us Researcher Workbench, click on the link inside and then reload this page.</span>
           </div>
         </div>
@@ -30,7 +48,6 @@
       <div class="alert alert-success" *ngIf="profile.emailVerificationStatus === 'subscribed'">
         <div class="alert-items">
           <div class="alert-item static">
-            <div class="alert-icon-wrapper"><clr-icon class="alert-icon" shape="check-circle"></clr-icon></div>
             <span class="alert-text">Your email address has been verified.</span>
           </div>
         </div>
@@ -38,10 +55,8 @@
     </div>
   </div>
 </div>
-<div class="field">
-  <h5>Username:</h5>
-  <div *ngIf="profileLoaded else loading" class="data">{{profile.username}}</div>
-</div>
+<h5>Username:</h5>
+<div *ngIf="profileLoaded else loading">{{profile.username}}</div>
 <ng-template #loading>
   <div class="loading-area">
     <span class="loading-text">Loading...</span>
@@ -90,22 +105,24 @@
       </div>
     </div>
   </ng-template>
-</div>
-<div *ngIf="profileLoaded">
-  <button class="btn" (click)="submitTermsOfService()"
-      [disabled]="email===null || this.profile.termsOfServiceCompletionTime !== null">
-    Submit Terms of Service
-  </button>
-</div>
-<div *ngIf="profileLoaded">
-  <button class="btn" (click)="completeEthicsTraining()"
-      [disabled]="email===null || this.profile.ethicsTrainingCompletionTime !== null">
-    Complete Ethics Training
-  </button>
-</div>
-<div *ngIf="profileLoaded">
-  <button class="btn" (click)="submitDemographicSurvey()"
-      [disabled]="email===null || this.profile.demographicSurveyCompletionTime !== null">
-    Submit Demographics Survey
-  </button>
+  <div *ngIf="profileLoaded">
+    <button class="btn" (click)="submitTermsOfService()"
+        [disabled]="email===null || this.profile.termsOfServiceCompletionTime !== null">
+      Submit Terms of Service
+    </button>
+  </div>
+  <div *ngIf="profileLoaded">
+    <button class="btn" (click)="completeEthicsTraining()"
+        [disabled]="email===null || this.profile.ethicsTrainingCompletionTime !== null">
+      Complete Ethics Training
+    </button>
+  </div>
+  <div *ngIf="profileLoaded">
+    <button class="btn" (click)="submitDemographicSurvey()"
+        [disabled]="email===null || this.profile.demographicSurveyCompletionTime !== null">
+      Submit Demographics Survey
+    </button>
+  </div>
+  <button *ngIf="editing" (click)="submitChanges()" class="btn">Update Profile</button>
+  <button *ngIf="editing" (click)="editing=false" class="btn">Cancel</button>
 </div>
